+++ core/src/main/scala/peloton/http/ActorSystemServer.scala
@@ -1,10 +1,10 @@
 package peloton.http
 
-import peloton.actor.ActorSystem
 import peloton.actor.Actor.CanAsk
 import peloton.actor.Actor.canAsk
-import peloton.http.Codecs.given
+import peloton.actor.ActorSystem
 import peloton.actor.ActorRef
+import peloton.http.Codecs.given
 import peloton.utils.Kryo
 
 import cats.effect.IO