+++ core/src/main/scala/peloton/http/Codecs.scala
@@ -1,6 +1,7 @@
 package peloton.http
 
 import cats.syntax.either.*
+
 import io.circe.Decoder
 import io.circe.Encoder
 import io.circe.Json
@@ -10,11 +11,10 @@ import scala.concurrent.duration.Duration
 
 object Codecs:
   
-  given Decoder[FiniteDuration] = Decoder.decodeLong.emap { dur => 
-    Either.catchNonFatal {
-      Duration.fromNanos(dur)
-    }.leftMap(_ => "Decoder[FiniteDuration]")
-  }
+  given Decoder[FiniteDuration] = Decoder.decodeLong.emap: dur => 
+    Either
+      .catchNonFatal(Duration.fromNanos(dur))
+      .leftMap(_ => "Decoder[FiniteDuration]")
 
   given Encoder[FiniteDuration] with
     final def apply(dur: FiniteDuration): Json = Json.fromLong(dur.toNanos)