+++ core/src/main/scala/peloton/persistence/PersistenceId.scala
@@ -1,5 +1,6 @@
 package peloton.persistence
 
+
 opaque type PersistenceId = String
 
 extension (id: PersistenceId)