+++ core/src/main/scala/peloton/persistence/Retention.scala
@@ -1,5 +1,6 @@
 package peloton.persistence
 
+
 /**
   * Retention parameters. 
   * 
@@ -7,7 +8,14 @@ package peloton.persistence
   * a new snapshot.
   *
   * @param purgeOnSnapshot
+  *   Flag. If set to `true`, the event store will keep a maximum of `snapshotsToKeep` snapshots that 
+  *   exist in the event store (including the one just created) and purge all events and snapshots that 
+  *   are older. If set to `false`, no purging will be performed.
   * @param snapshotsToKeep
+  *   Number of snapshots to keep when purging is requested (`purgeOnSnapshot` is set to `true`). 
+  * 
+  *   **Attention!** This number includes the snapshot just created, so if this parameter is set to 0, 
+  *   all snapshots (including the the one just created!) will be deleted.
   */
 final case class Retention(
   purgeOnSnapshot: Boolean = false,