+++ core/src/test/scala/peloton/actors/CalculatorActor.scala
@@ -3,16 +3,16 @@ package peloton.actors
 import cats.effect.IO
 
 import peloton.actor.Actor
+import peloton.actor.Actor.{CanAsk, canAsk}
+import peloton.actor.ActorContext
 import peloton.actor.ActorRef
-import peloton.actor.Actor.*
 import peloton.actor.ActorSystem
-import peloton.actor.EventAction
 import peloton.persistence.PersistenceId
 import peloton.persistence.EventStore
+import peloton.persistence.EventAction
 import peloton.persistence.PayloadCodec
 import peloton.persistence.KryoPayloadCodec
-import peloton.actor.ActorContext
-import peloton.actor.SnapshotPredicate
+import peloton.persistence.SnapshotPredicate
 
 /**
   * An event sourced actor with simple calculation operations